+++ R/qcMWRfre.R
@@ -1,12 +1,13 @@
 #' Run quality control frequency checks for water quality monitoring results
 #'
 #' @param res character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}
+#' @param acc character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}
 #' @param frecom character string of path to the data quality objectives file for frequency and completeness or \code{data.frame} returned by \code{\link{readMWRfrecom}}
 #' @param fset optional list of inputs with elements named \code{res}, \code{acc}, \code{frecom}, \code{sit}, or \code{wqx} overrides the other arguments
 #' @param runchk  logical to run data checks with \code{\link{checkMWRresults}} and \code{\link{checkMWRfrecom}}, applies only if \code{res} or \code{frecom} are file paths
 #' @param warn logical to return warnings to the console (default)
 #'
-#' @details The function can be used with inputs as paths to the relevant files or as data frames returned by \code{\link{readMWRresults}} and \code{\link{readMWRfrecom}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}.  In the latter case, downstream analyses may not work if data are formatted incorrectly. For convenience, a named list with the input arguments as paths or data frames can be passed to the \code{fset} argument instead. See the help file for \code{\link{utilMWRinput}}.
+#' @details The function can be used with inputs as paths to the relevant files or as data frames returned by \code{\link{readMWRresults}}, \code{\link{readMWRacc}}, and \code{\link{readMWRfrecom}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}.  In the latter case, downstream analyses may not work if data are formatted incorrectly. For convenience, a named list with the input arguments as paths or data frames can be passed to the \code{fset} argument instead. See the help file for \code{\link{utilMWRinput}}.
 #' 
 #' Note that frequency is only evaluated on parameters in the \code{Parameter} column in the data quality objectives frequency and completeness file.  A warning is returned if there are parameters in \code{Parameter} in the frequency and completeness file that are not in \code{Characteristic Name} in the results file. 
 #' 
@@ -22,12 +23,15 @@
 #' 
 #' # results path
 #' respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')
+#'
+#' # dqo accuracy data path
+#' accpth <- system.file('extdata/ExampleDQOAccuracy.xlsx', package = 'MassWateR')
 #' 
 #' # frequency and completeness path
 #' frecompth <- system.file('extdata/ExampleDQOFrequencyCompleteness.xlsx', 
 #'      package = 'MassWateR')
 #' 
-#' qcMWRfre(res = respth, frecom = frecompth)
+#' qcMWRfre(res = respth, acc = accpth, frecom = frecompth)
 #' 
 #' ##
 #' # using data frames
@@ -35,19 +39,24 @@
 #' # results data
 #' resdat <- readMWRresults(respth)
 #' 
+#' # accuracy data
+#' accdat <- readMWRacc(accpth)
+#' 
 #' # frequency and completeness data
 #' frecomdat <- readMWRfrecom(frecompth)
 #' 
-#' qcMWRfre(res = resdat, frecom = frecomdat)
-#' 
-qcMWRfre <- function(res = NULL, frecom = NULL, fset = NULL, runchk = TRUE, warn = TRUE){
+#' qcMWRfre(res = resdat, acc = accdat, frecom = frecomdat)
+qcMWRfre <- function(res = NULL, acc = NULL, frecom = NULL, fset = NULL, runchk = TRUE, warn = TRUE){
   
   utilMWRinputcheck(mget(ls()))
   
+  colsym <- c('<=', '<', '>=', '>', '\u00b1', '\u2265', '\u2264', '%', 'AQL', 'BDL', 'log', 'all')
+  
   ##
   # get user inputs
-  inp <- utilMWRinput(res = res, frecom = frecom, fset = fset, runchk = runchk, warn = warn)
+  inp <- utilMWRinput(res = res, acc = acc, frecom = frecom, fset = fset, runchk = runchk, warn = warn)
   resdat <- inp$resdat
+  accdat <- inp$accdat
   frecomdat <- inp$frecomdat
   
   ##
@@ -77,16 +86,12 @@ qcMWRfre <- function(res = NULL, frecom = NULL, fset = NULL, runchk = TRUE, warn
   # run completeness checks
   for(prm in prms){
 
-    # subset dqo data
-    frecomdattmp <- frecomdat %>% 
-      dplyr::filter(Parameter == prm)
+    # subset results data, filter by value range
+    resdattmp <- utilMWRfre(resdat, param = prm, accdat) 
     
-    # subset results data
-    resdattmp <- resdat %>% 
-      dplyr::filter(`Characteristic Name` == prm)
-
-    # total obs
-    ntot <- resdattmp %>% 
+    # total obs, includes all regardless of value range or dqo
+    ntot <- resdat %>% 
+      dplyr::filter(`Characteristic Name` == prm) %>% 
       dplyr::filter(`Activity Type` %in% c('Sample-Routine', 'Field Msr/Obs')) %>%
       nrow()
     
@@ -94,24 +99,37 @@ qcMWRfre <- function(res = NULL, frecom = NULL, fset = NULL, runchk = TRUE, warn
     fielddup <- resdattmp %>% 
       dplyr::filter(`Activity Type` %in% c('Sample-Routine', 'Field Msr/Obs')) %>% 
       dplyr::filter(!is.na(`QC Reference Value`)) %>% 
+      dplyr::filter(!is.na(`Field Duplicate`)) %>% 
       nrow()
-    
+
     # lab duplicates
     acts <- c('Quality Control Sample-Lab Duplicate', 'Quality Control-Meter Lab Duplicate')
-    labdup <- sum(resdattmp$`Activity Type` %in% acts)
+    labdup <- ifelse(
+      any(!is.na(resdattmp$`Lab Duplicate`)), 
+      sum(resdattmp$`Activity Type` %in% acts),
+      0)
     
     # field blank
     acts <- 'Quality Control Sample-Field Blank'
-    fieldblnk <- sum(resdattmp$`Activity Type` %in% acts)
+    fieldblnk <- ifelse(
+      any(!is.na(resdattmp$`Field Blank`)), 
+      sum(resdattmp$`Activity Type` %in% acts),
+      0)
     
     # lab blank
     acts <- c('Quality Control Sample-Lab Blank', 'Quality Control-Meter Lab Blank')
-    labblnk <- sum(resdattmp$`Activity Type` %in% acts)
+    labblnk <- ifelse(
+      any(!is.na(resdattmp$`Lab Blank`)), 
+      sum(resdattmp$`Activity Type` %in% acts), 
+      0)
 
     # lab spikes or instrument checks
     acts <- c('Quality Control Sample-Lab Spike', 'Quality Control-Calibration Check')
-    spikesinstchks <- sum(resdattmp$`Activity Type` %in% acts, na.rm = TRUE)
-    
+    spikesinstchks <- ifelse(
+      any(!is.na(resdattmp$`Spike/Check Accuracy`)), 
+      sum(resdattmp$`Activity Type` %in% acts, na.rm = TRUE), 
+      0)
+
     # compile results
     res <- tibble::tibble(
       Parameter = prm, 
@@ -134,15 +152,16 @@ qcMWRfre <- function(res = NULL, frecom = NULL, fset = NULL, runchk = TRUE, warn
       `Instrument Check` = `Spike/Check Accuracy`
     ) %>% 
     tidyr::pivot_longer(cols = -dplyr::matches('Parameter'), names_to = 'check', values_to = 'standard')
-  
+
   # summary results long format
   resall <- resall %>% 
     tidyr::pivot_longer(cols = -dplyr::matches('^Parameter$|^obs$'), names_to = 'check', values_to = 'count')
-  
+
   # combine and create summaries
   out <- resall %>% 
     dplyr::left_join(frecomdat, by = c('Parameter', 'check')) %>% 
     dplyr::mutate(
+      count = as.integer(count),
       percent = ifelse(
         !is.na(standard), 100 * count / obs,
         NA_real_