+++ tests/testthat/test-checkMWRacc.R
@@ -10,19 +10,50 @@ test_that("Checking required column names are present", {
   expect_error(checkMWRacc(chk))
 })
 
-test_that("Checking column types", {
+test_that("Checking column types, including NA", {
   chk <- accdatchk
-  chk$`MDL` <- as.character(chk$`MDL`)
-  chk$`UQL`[5] <- 'a'
-  chk$`Spike/Check Accuracy` <- suppressWarnings(as.numeric(chk$`Spike/Check Accuracy`))
-  expect_error(checkMWRacc(chk), regexp = '\tChecking column types...\n\tIncorrect column type found in columns: UQL-character, Spike/Check Accuracy-numeric', fixed = T)
+  chk$`UQL` <- NA
+  chk$`MDL` <- 'a'
+  chk$`Spike/Check Accuracy` <- 5
+  chk$`Lab Duplicate` <- NA
+  expect_error(checkMWRacc(chk), regexp = '\tChecking column types...\n\tIncorrect column type found in columns: MDL-should be numeric, Spike/Check Accuracy-should be character', fixed = T)
+})
+
+test_that("Checking column types with NA values", {
+  chk <- accdatchk
+  chk$`UQL` <- NA
+  chk$`Spike/Check Accuracy` <- NA
+  result <- suppressWarnings(checkMWRacc(chk))
+  expect_s3_class(result, 'tbl_df')
 })
 
 test_that("Checking for text other than <=, \u2264, <, >=, \u2265, >, \u00b1, %, AQL, BQL, log, or all", {
   chk <- accdatchk
-  chk$`Value Range`[4] <- '+'
+  chk$`Value Range`[4] <- 'b'
   chk$`Field Duplicate` <- 'alll'
   expect_error(checkMWRacc(chk), regexp = 'Unrecognized text in columns: Value Range, Field Duplicate', fixed = T)
+})
+
+test_that("Checking for more than two rows per parameter", {
+  chk <- accdatchk
+  chk <- bind_rows(accdatchk, accdatchk[8,], accdatchk[6,])
+  expect_error(checkMWRacc(chk), regexp = 'More than two rows: Sp Conductance, TP', fixed = T)
+})
+
+test_that("Checking overlap in value range", {
+  chk <- accdatchk
+  chk$`Value Range`[11] <- '<60'
+  chk$`Value Range`[3] <- '<= 4'
+  chk$`Value Range`[7] <- '<= 0.06'
+  expect_error(checkMWRacc(chk), regexp = 'Overlap in value range: DO, E.coli, TP', fixed = T)
+})
+
+test_that("Checking gap in value range", {
+  chk <- accdatchk
+  chk$`Value Range`[12] <- '>60'
+  chk$`Value Range`[10] <- '>12'
+  chk$`Value Range`[3] <- '< 1'
+  expect_warning(checkMWRacc(chk), regexp = 'Gap in value range in DQO accuracy file: Ammonia, DO, E.coli', fixed = T)
 })
 
 test_that("Checking missing entries in uom", {
@@ -55,7 +86,7 @@ test_that("Checking incorrect unit type per parameter", {
 
 test_that("Checking empty column", {
   chk <- accdatchk
-  chk[[4]] <- NA
+  chk[[7]] <- NA
   expect_warning(checkMWRacc(chk))
 })