+++ tests/testthat/test-qcMWRacc.R
@@ -10,6 +10,14 @@ test_that("Checking warning if parameters in acc missing from res", {
   expect_warning(qcMWRacc(resdatchk, accpth, frecompth, runchk = F))
 })
 
+test_that("Checking warning if gap in value range", {
+  accdatchk <- readMWRacc(accpth, runchk = F, warn = F)
+  accdatchk$`Value Range`[12] <- '>60'
+  accdatchk$`Value Range`[10] <- '>12'
+  accdatchk$`Value Range`[3] <- '< 1'
+  expect_warning(qcMWRacc(respth, accdatchk, frecompth, runchk = F), regexp = 'Gap in value range in DQO accuracy file: Ammonia, DO, E.coli', fixed = T)
+})
+
 test_that("Checking error if unit mismatch between results and accuracy file", {
   resdatchk <- resdat %>% 
     mutate(
@@ -51,4 +59,15 @@ test_that("Checking empty data frame in list output", {
     )
   result <- qcMWRacc(respth, accchk, frecomchk, runchk = F, warn = F)
   expect_type(result, 'list')
+})
+
+test_that("Checking error if upper value range for spikes/checks is not percent if unit is percent", {
+  accdatchk <- accdat %>% 
+    mutate(
+      `Spike/Check Accuracy` = case_when(
+        Parameter == 'TP' & grepl('>', `Value Range`) ~ '<= 15', 
+        TRUE ~ `Spike/Check Accuracy`
+      )
+    )
+  expect_error(qcMWRacc(resdat, accdatchk, frecomdat, warn = F), 'Lab Spikes / Instrument Checks with units as % must have DQO accuracy as % for upper value range: TP')
 })