+++ NEWS.md
@@ -1,12 +1,28 @@
+# MassWateR 2.1.2
+
+* Warning is now returned if a column for a QC check includes all `na` entries on import of the DQO accuracy file with `readMWRacc()`
+* `readMWRacc()` no longer uses `dplyr::na_if()` on all columns, only for numeric, to correctly identify columns that have all `na` entries for a QC check
+* Sorting behavior for `dplyr::arrange()` reverted to `.locale = 'en'` to ignore case
+* `utilMWRfre()` function added to prep results data for frequency checks, similar to existing `utilMWRlimits()` function
+* Correct number of data records is now reported by `tabMWRfre()` following value range filtering
+* `tabMWRfre()` and `qcMWRfre()` now require the DQO accuracy file as input to identify appropriate ranges to check for each parameter using the value range column
+* Added `utilMWRvaluerange()` function to check for `na`, gaps, and overlap in the value range column on import of the DQO accuracy file with `readMWRacc()`
+* Error is now returned if overlapping value ranges are present in the DQO accuracy file
+* Warning now returned if gaps are present in the value ranges for a parameter in the DQO accuracy file
+* Fix to `checkMWRacc()` to convert MDL and UQL columns in DQO accuracy as numeric following import as text
+* Fixed bug to evaluate lab spike QC checks as the absolute difference, was previously a relative difference
+* New error message if the upper value range in the DQO accuracy file is not a percent value for lab spike QC checks with units as percentage
+* Lab spikes entered as a percent measure are now evaluated against the correct data quality objective using the upper value range
+* Better error and warning messages for `tabMWRacc()` for incorrect and required data for individual QC checks
+
 # MassWateR 2.1.1
 
 * Fix to incorrect secondary label for Enterococcus in `thresholdMWR`
 * Corrected column name to `Record ID User Supplied` in WQX output from `tabMWRwqx()`
 * Fix for `tabMWRwqx()` output if no entries in `QC Reference Value` column in the results file, this previously resulted in an error
 * `formMWRresults()` no longer imports dplyr and tidyr in Roxygen documentation
-* `formMWRresults()` now handles time and text inputs for `Activity Start Date` from Excel, all times are correctly formatted as HH:MM in 24 hour time.  This includes more robust testing of inputs in `text-formMWRresults.R`.
+* `formMWRresults()` now handles time and text inputs for `Activity Start Time` from Excel, all times are correctly formatted as HH:MM in 24 hour time.  This includes more robust testing of inputs in `text-formMWRresults.R`.
 * Added check for data quality objectives for accuracy that returns an error if `na` entries are included in the `Value Range` column, should be `all`
-* Fix to date output for the `Activity Start Date` column returned by `tabMWRwqx()` to be `YYYY-MM-DD` character, empty time was previously included as date time object
 
 # MassWateR 2.1.0